+++ packages/orq-rc/src/models/operations/retrievedatasource.ts
@@ -172,7 +172,7 @@ export const RetrieveDatasourceResponseBody$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("01K6JQQJ0M12F6NB7KMJ2NT3HD"),
+  _id: z.string().default("01K6MDVPQ5E1Y7YC7MB3GE5RJW"),
   display_name: z.string(),
   description: z.string().optional(),
   status: RetrieveDatasourceStatus$inboundSchema,
@@ -216,7 +216,7 @@ export const RetrieveDatasourceResponseBody$outboundSchema: z.ZodType<
   z.ZodTypeDef,
   RetrieveDatasourceResponseBody
 > = z.object({
-  id: z.string().default("01K6JQQJ0M12F6NB7KMJ2NT3HD"),
+  id: z.string().default("01K6MDVPQ5E1Y7YC7MB3GE5RJW"),
   displayName: z.string(),
   description: z.string().optional(),
   status: RetrieveDatasourceStatus$outboundSchema,